requirements.txt
scikit-learn
numpy
pandas
faiss-cpu
pyyaml
requests
joblib
matplotlib
seaborn